requirements.txt
aiosqlite==0.11.0
asyncpg==0.20.1
asyncssh==2.2.0
bcrypt==3.1.7
cffi==1.14.0
ciso8601==2.1.3
click==7.1.1
cryptography==2.8
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.20
PyNaCl==1.3.0
PyPika==0.36.0
python-multipart==0.0.5
six==1.14.0
starlette==0.13.2
tortoise-orm==0.16.3
typing-extensions==3.7.4.1
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
